+++ py3/classes/sysflow/reader.py
@@ -223,7 +223,7 @@ class FlattenedSFReader(SFReader):
                     container = None
                     if rec.containerId is not None:
                         if not rec.containerId in self.conts:
-                            print("ERROR: Cannot find container object for record. This should not happen.")
+                            print("WARN: Cannot find container object for record. This should not happen.")
                         else:
                             container = self.conts[rec.containerId]
                     return (ObjectTypes.PROC, self.header, container, None, rec, None, None, None)
@@ -234,7 +234,7 @@ class FlattenedSFReader(SFReader):
                     container = None
                     if rec.containerId is not None:
                         if not rec.containerId in self.conts:
-                            print("ERROR: Cannot find container object for record. This should not happen.")
+                            print("WARN: Cannot find container object for record. This should not happen.")
                         else:
                             container = self.conts[rec.containerId]
                     return (ObjectTypes.FILE, self.header, container, None, None, (rec, None), None, None)
@@ -248,27 +248,27 @@ class FlattenedSFReader(SFReader):
                 evt = None
                 flow = None
                 if not procOID in self.processes:
-                    print("ERROR: Cannot find process object for record. This should not happen.")
+                    print("WARN: Cannot find process object for record. This should not happen.")
                 else:
                     proc = self.processes[procOID]
                     pproc = self.getProcess(proc.poid) if proc.poid is not None else None
                 if proc is not None:
                     if proc.containerId is not None:
                         if not proc.containerId in self.conts:
-                            print("ERROR: Cannot find container object for record. This should not happen.")
+                            print("WARN: Cannot find container object for record. This should not happen.")
                         else:
                             container = self.conts[proc.containerId]
                 if objtype == ObjectTypes.FILE_EVT:
                     fileOID = rec.fileOID
                     evt = rec
                     if not fileOID in self.files:
-                        print("ERROR: Cannot find file object for record. This should not happen.")
+                        print("WARN: Cannot find file object for record. This should not happen.")
                     else:
                         file1 = self.files[fileOID]
                     fileOID2 = rec.newFileOID
                     if fileOID2 is not None:
                         if not fileOID2 in self.files:
-                            print("ERROR: Cannot find file object for record. This should not happen.")
+                            print("WARN: Cannot find file object for record. This should not happen.")
                         else:
                             file2 = self.files[fileOID2]
 
@@ -276,7 +276,7 @@ class FlattenedSFReader(SFReader):
                     fileOID = rec.fileOID
                     flow = rec
                     if not fileOID in self.files:
-                        print("ERROR: Cannot find file object for record. This should not happen.")
+                        print("WARN: Cannot find file object for record. This should not happen.")
                     else:
                         file1 = self.files[fileOID]
                 elif objtype == ObjectTypes.PROC_EVT: